# brain-backend/api-ref/brain_openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Brain API",
        "version": "0.1.0"
    },
    "paths": {
...
        "/mv-servers": {
            "get": {
                "tags": [
                    "mv200"
                ],
                "summary": "Get All Mv Servers",
                "description": "Get all MV servers list",
                "operationId": "get_all_mv_servers_mv_servers_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/MVServer"
                                    },
                                    "type": "array",
                                    "title": "Response Get All Mv Servers Mv Servers Get"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "mv200"
                ],
                "summary": "Create Mv Server",
                "description": "Create a new MV server",
                "operationId": "create_mv_server_mv_servers_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MVServerCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MVServer"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/mv-servers/{server_id}": {
            "get": {
                "tags": [
                    "mv200"
                ],
                "summary": "Get Mv Server",
                "description": "Get specific MV server by ID",
                "operationId": "get_mv_server_mv_servers__server_id__get",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Server Id"
                        },
                        "name": "server_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MVServer"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "mv200"
                ],
                "summary": "Update Mv Server",
                "description": "Update MV server information by ID",
                "operationId": "update_mv_server_mv_servers__server_id__put",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Server Id"
                        },
                        "name": "server_id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MVServerUpdate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MVServer"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "mv200"
                ],
                "summary": "Delete Mv Server",
                "description": "Delete MV server by ID",
                "operationId": "delete_mv_server_mv_servers__server_id__delete",
                "parameters": [
                    {
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Server Id"
                        },
                        "name": "server_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "MVServer": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id",
                        "description": "Unique identifier for the MV server"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "ip_address": {
                        "type": "string",
                        "title": "Ip Address"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "host_ip": {
                        "type": "string",
                        "title": "Host Ip"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "name",
                    "ip_address"
                ],
                "title": "MVServer"
            },
            "MVServerCreate": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "ip_address": {
                        "type": "string",
                        "title": "Ip Address"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "host_ip": {
                        "type": "string",
                        "title": "Host Ip"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "ip_address"
                ],
                "title": "MVServerCreate"
            },
            "MVServerUpdate": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "ip_address": {
                        "type": "string",
                        "title": "Ip Address"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "host_ip": {
                        "type": "string",
                        "title": "Host Ip"
                    }
                },
                "type": "object",
                "title": "MVServerUpdate"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "login"
                    }
                }
            }
        }
    }
}